TypeScriptへの移行に伴い、index.jsをindex.tsに置き換え、.gitignoreにindex.jsを追加。Dockerfileを更新し、TypeScriptのインストールとコンパイルを追加。package.jsonにビルドスクリプトとdevDependenciesを追加。tsconfig.jsonを新規作成。
